Fixed regression in section links introduced in v1.10.
This was introduced in an effort to make the html output XHTML 1.0 STRICT compliant because the standard does not allow / characters in anchors.

However, the / characters were changed to . in the anchors but not in the links.  For now revert the anchors to / so further though can be given to this issue.