Fix difference in cipher type reporting missed in 8304d452.
The C code can't get the cipher type from the storage object because the C storage object does not have encryption baked in like the Perl code does.

Instead, check backup.info to see if encryption is enabled.  This will need to rethought if another cipher type is added but for now it works fine.